refactor: move and rename core components
