--- src/core/parser/index.ts
@@ -1,1 +1,0 @@
-export { ParserWorkerManager as default } from "./manager"
