+++ src/core/parsing/traverse.ts
