+++ src/core/parsing/worker.ts
